Reorganize some files after 6pak review
