+++ Reactor/Reactor.csproj
@@ -14,7 +14,7 @@
 
     <ItemGroup>
         <PackageReference Include="Microsoft.SourceLink.GitHub" Version="1.1.1" PrivateAssets="All" />
-        <Compile Include="..\Reactor.Networking.Shared\**\*.cs" LinkBase="Reactor.Networking.Shared" Visible="false" />
+        <Compile Include="..\Reactor.Networking.Shared\**\*.cs" LinkBase="Networking" Visible="false" />
     </ItemGroup>
 
     <ItemGroup>
